+++ rio_tiler/colormap.py
@@ -2,7 +2,6 @@
 
 import os
 import pathlib
-from copy import deepcopy
 from typing import Dict, List, Sequence, Tuple, Union
 
 import attr
@@ -75,7 +74,7 @@ def apply_cmap(
         colormap (dict): GDAL RGBA Color Table dictionary.
 
     Returns:
-        tuple: Data (numpy.ndarray) and Alpha band (numpy.ndarray).
+        tuple: Data (numpy.ndarray) and Mask (numpy.ndarray) values.
 
     Raises:
         InvalidFormat: If data is not a 1 band dataset (1, col, row).
@@ -138,11 +137,13 @@ class ColorMaps:
     """Default Colormaps holder.
 
     Args:
-        data (dict): colormaps.
+        data (dict): colormaps. Defaults to `rio_tiler.colormap.DEFAULTS_CMAPS`.
 
     """
 
-    data: Dict[str, Union[str, numpy.array]] = attr.ib()
+    data: Dict[str, Union[str, numpy.array]] = attr.ib(
+        default=attr.Factory(lambda: DEFAULTS_CMAPS)
+    )
 
     def get(self, name: str) -> Dict:
         """Fetch a colormap.
@@ -188,6 +189,7 @@ class ColorMaps:
             >>> cmap = cmap.register({"acmap": {0: [0, 0, 0, 0]}})
 
             >>> cmap = cmap.register({"acmap": "acmap.npy"})
+
         """
         for name, cmap in custom_cmap.items():
             if not overwrite and name in self.data:
@@ -198,4 +200,4 @@ class ColorMaps:
         return ColorMaps({**self.data, **custom_cmap})
 
 
-cmap = ColorMaps(deepcopy(DEFAULTS_CMAPS))  # noqa
+cmap = ColorMaps()  # noqa